fix: mentors check 2
